+++ selectrum.el
@@ -1093,38 +1093,38 @@ will be set to `selectrum-num-candidates-displayed' if
       (window-resize
        window (- dheight wheight) nil nil 'pixelwise))))
 
-(defun selectrum--ensure-single-lines (candidates)
+(defun selectrum--ensure-single-lines (candidates settings)
   "Return list of single-line CANDIDATES.
 Multi-line candidates are merged into a single line. The resulting
 single-line candidates are then shortened by replacing repeated
 whitespace and maybe truncating the result.
 
 The specific details of the formatting are determined by
-`selectrum-multiline-display-settings'."
+SETTINGS, see `selectrum-multiline-display-settings'."
   (let* ((single/lines ())
 
          ;; The formatting settings are the same for all multi-line
          ;; candidates, and so only need to be gotten once from
-         ;; `selectrum-multiline-display-settings'.
+         ;; `settings'.
          ;;
          ;; - Matching lines
          (match/transformation
-          (alist-get 'match selectrum-multiline-display-settings))
+          (alist-get 'match settings))
          (match/display (car match/transformation))
          (match/face (cadr match/transformation))
          ;; - Truncated candidate
          (truncation/transformation
-          (alist-get 'truncation selectrum-multiline-display-settings))
+          (alist-get 'truncation settings))
          (truncation/display (car truncation/transformation))
          (truncation/face (cadr truncation/transformation))
          ;; - Newlines
          (newline/transformation
-          (alist-get 'newline selectrum-multiline-display-settings))
+          (alist-get 'newline settings))
          (newline/display (car newline/transformation))
          (newline/face (cadr newline/transformation))
          ;; - Repeated whitespace
          (whitespace/transformation
-          (alist-get 'whitespace selectrum-multiline-display-settings))
+          (alist-get 'whitespace settings))
          (whitespace/display (car whitespace/transformation))
          (whitespace/face (cadr whitespace/transformation)))
 
@@ -1132,19 +1132,24 @@ The specific details of the formatting are determined by
       (if (string-match-p "\n" cand)
           (let* ((lines (split-string cand "\n"))
                  (len (length lines))
-                 (fmatch (car (funcall
-                               selectrum-refine-candidates-function
-                               (minibuffer-contents)
-                               lines)))
+                 (input (minibuffer-contents))
+                 (fmatch (if (string-empty-p input)
+                             (with-temp-buffer
+                               (insert cand)
+                               (goto-char (point-min))
+                               (skip-chars-forward " \t\n")
+                               (buffer-substring (line-beginning-position)
+                                                 (line-end-position)))
+                           (car (funcall
+                                 selectrum-refine-candidates-function
+                                 input
+                                 lines))))
                  (match
                   (propertize
                    (replace-regexp-in-string
                     "[ \t][ \t]+"
                     (propertize whitespace/display 'face whitespace/face)
-                    (if (string-empty-p (minibuffer-contents))
-                        ""
-                      ;; Show first matched line.
-                      (or fmatch "")) 'fixed-case 'literal)
+                    (or fmatch "") 'fixed-case 'literal)
                    'selectrum-candidate-display-prefix
                    (propertize (format "(%d lines)" len)
                                'face newline/face)))
@@ -1299,9 +1304,11 @@ defaults to `minibuffer-completion-table'. PRED defaults to
                                                  :annotf annotf
                                                  :docsigf docsigf))
                            (t candidates)))
-         (lines (selectrum--ensure-single-lines candidates)))
+         (extend selectrum-extend-current-candidate-highlight)
+         (show-indices selectrum-show-indices)
+         (margin-padding selectrum-right-margin-padding))
     (with-temp-buffer
-      (dolist (candidate lines)
+      (dolist (candidate candidates)
         (let* ((prefix (get-text-property
                         0 'selectrum-candidate-display-prefix
                         candidate))
@@ -1343,9 +1350,9 @@ defaults to `minibuffer-completion-table'. PRED defaults to
             (when annot-fun
               (funcall annot-fun prefix suffix right-margin)))
           (insert "\n")
-          (when selectrum-show-indices
-            (let* ((display-fn (if (functionp selectrum-show-indices)
-                                   selectrum-show-indices
+          (when show-indices
+            (let* ((display-fn (if (functionp show-indices)
+                                   show-indices
                                  (lambda (i) (format "%2d " i))))
                    (curr-index (substring-no-properties
                                 (funcall display-fn (1+ index)))))
@@ -1364,12 +1371,12 @@ defaults to `minibuffer-completion-table'. PRED defaults to
                'display
                `(space :align-to (- right-fringe
                                     ,(string-width right-margin)
-                                    selectrum-right-margin-padding)))
+                                    ,margin-padding)))
               (if formatting-current-candidate
                   (selectrum--add-face
                    right-margin'selectrum-current-candidate)
                 right-margin))))
-           ((and selectrum-extend-current-candidate-highlight
+           ((and extend
                  formatting-current-candidate)
             (insert
              (propertize
@@ -1377,7 +1384,7 @@ defaults to `minibuffer-completion-table'. PRED defaults to
               'face 'selectrum-current-candidate
               'display
               `(space :align-to (- right-fringe
-                                   selectrum-right-margin-padding)))))))
+                                   ,margin-padding)))))))
         (cl-incf index))
       (split-string (buffer-string) "\n" t))))