+++ CHANGELOG.md
@@ -64,9 +64,13 @@ The format is based on [Keep a Changelog].
   prompt was reinserted in place so it did not have any useful effect
   ([#263]).
 * Multiline candidates are now prefixed with the number of newlines
-  they contain ([#266], [#302]).
+  they contain ([#266], [#302], [#318]).
 
 ### Bugs fixed
+* `selectrum-extend-current-candidate-highlight`,
+  `selectrum-show-indices`, `selectrum-right-margin-padding` and
+  `selectrum-multiline-display-settings` wouldn't use the local
+  session bindings if there were any, which has been fixed ([#317]).
 * `selectrum-insert-current-candidate` would duplicate the prompt for
   `completing-read-multiple` when the prompt was selected, which has
   been fixed. The behavior is now like in `completing-read` ([#296]).
@@ -148,6 +152,8 @@ The format is based on [Keep a Changelog].
 [#309]: https://github.com/raxod502/selectrum/pull/309
 [#312]: https://github.com/raxod502/selectrum/issues/312
 [#316]: https://github.com/raxod502/selectrum/pull/316
+[#317]: https://github.com/raxod502/selectrum/pull/317
+[#318]: https://github.com/raxod502/selectrum/pull/318
 
 ## 3.0 (released 2020-10-20)
 ### Breaking changes